+++ Sources/StreamFeeds/StateLayer/FeedQuery.swift
@@ -161,7 +161,7 @@ extension FeedQuery {
             activitySelectorOptions: activitySelectorOptions,
             data: data,
             externalRanking: externalRanking,
-            filter: activityFilter?.toRawJSON(),
+            filter: activityFilter?.toRawJSONDictionary(),
             followersPagination: followerLimit.flatMap { PagerRequest(limit: $0) },
             followingPagination: followingLimit.flatMap { PagerRequest(limit: $0) },
             interestWeights: interestWeights,